removed warning
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@5915 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/gridpart/codimindexset.hh
@@ -75,7 +75,7 @@ private:
   const int myCodim_; 
 
   // actual number of holes 
-  int numberHoles_;
+  size_t numberHoles_;
 
 public:
   //! Constructor taking memory factor (default = 1.1)
